chore: refactor to separate each of the program creation types
