+++ packages/typescript-estree/src/parser.ts
@@ -1,22 +1,16 @@
-import debug from 'debug';
-import path from 'path';
 import semver from 'semver';
 import * as ts from 'typescript'; // leave this as * as ts so people using util package don't need syntheticDefaultImports
 import { sync as globSync } from 'glob';
 import isGlob from 'is-glob';
 import { astConverter } from './ast-converter';
 import { convertError } from './convert';
-import { firstDefined } from './node-utils';
+import { createDefaultProgram } from './create-program/createDefaultProgram';
+import { createIsolatedProgram } from './create-program/createIsolatedProgram';
+import { createProjectProgram } from './create-program/createProjectProgram';
+import { createSourceFile } from './create-program/createSourceFile';
 import { Extra, TSESTreeOptions, ParserServices } from './parser-options';
 import { getFirstSemanticOrSyntacticError } from './semantic-or-syntactic-errors';
 import { TSESTree } from './ts-estree';
-import {
-  calculateProjectParserOptions,
-  createProgram,
-  defaultCompilerOptions,
-} from './tsconfig-parser';
-
-const log = debug('typescript-eslint:typescript-estree:parser');
 
 /**
  * This needs to be kept in sync with the top-level README.md in the
@@ -32,18 +26,6 @@ const isRunningSupportedTypeScriptVersion = semver.satisfies(
 let extra: Extra;
 let warnedAboutTSVersion = false;
 
-/**
- * Compute the filename based on the parser options.
- *
- * Even if jsx option is set in typescript compiler, filename still has to
- * contain .tsx file extension.
- *
- * @param options Parser options
- */
-function getFileName({ jsx }: { jsx?: boolean }): string {
-  return jsx ? 'estree.tsx' : 'estree.ts';
-}
-
 function enforceString(code: unknown): string {
   /**
    * Ensure the source code is a string
@@ -53,6 +35,44 @@ function enforceString(code: unknown): string {
   }
 
   return code;
+}
+
+interface ASTAndProgram {
+  ast: ts.SourceFile;
+  program: ts.Program | undefined;
+}
+
+/**
+ * @param code The code of the file being linted
+ * @param options The config object
+ * @param shouldProvideParserServices True iff the program should be attempted to be calculated from provided tsconfig files
+ * @returns Returns a source file and program corresponding to the linted code
+ */
+function getProgramAndAST(
+  code: string,
+  shouldProvideParserServices: boolean,
+  shouldCreateDefaultProgram: boolean,
+): ASTAndProgram | undefined {
+  return (
+    (shouldProvideParserServices &&
+      createProjectProgram(code, shouldCreateDefaultProgram, extra)) ||
+    (shouldProvideParserServices &&
+      shouldCreateDefaultProgram &&
+      createDefaultProgram(code, extra)) ||
+    createIsolatedProgram(code, extra)
+  );
+}
+
+/**
+ * Compute the filename based on the parser options.
+ *
+ * Even if jsx option is set in typescript compiler, filename still has to
+ * contain .tsx file extension.
+ *
+ * @param options Parser options
+ */
+function getFileName({ jsx }: { jsx?: boolean } = {}): string {
+  return jsx ? 'estree.tsx' : 'estree.ts';
 }
 
 /**
@@ -67,6 +87,7 @@ function resetExtra(): void {
     errorOnTypeScriptSyntacticAndSemanticIssues: false,
     errorOnUnknownASTType: false,
     extraFileExtensions: [],
+    filePath: getFileName(),
     jsx: false,
     loc: false,
     log: console.log, // eslint-disable-line no-console
@@ -81,173 +102,6 @@ function resetExtra(): void {
   };
 }
 
-interface ASTAndProgram {
-  ast: ts.SourceFile;
-  program: ts.Program | undefined;
-}
-
-/**
- * @param code The code of the file being linted
- * @param options The config object
- * @returns If found, returns the source file corresponding to the code and the containing program
- */
-function getASTFromProject(
-  code: string,
-  options: TSESTreeOptions,
-  createDefaultProgram: boolean,
-): ASTAndProgram | undefined {
-  log('Attempting to get AST from project(s) for: %s', options.filePath);
-
-  const filePath = options.filePath || getFileName(options);
-  const astAndProgram = firstDefined(
-    calculateProjectParserOptions(code, filePath, extra),
-    currentProgram => {
-      const ast = currentProgram.getSourceFile(filePath);
-      return ast && { ast, program: currentProgram };
-    },
-  );
-
-  if (!astAndProgram && !createDefaultProgram) {
-    // the file was either not matched within the tsconfig, or the extension wasn't expected
-    const errorLines = [
-      '"parserOptions.project" has been set for @typescript-eslint/parser.',
-      `The file does not match your project config: ${filePath}.`,
-    ];
-    let hasMatchedAnError = false;
-
-    const fileExtension = path.extname(filePath);
-    if (!['.ts', '.tsx', '.js', '.jsx'].includes(fileExtension)) {
-      const nonStandardExt = `The extension for the file (${fileExtension}) is non-standard`;
-      if (extra.extraFileExtensions && extra.extraFileExtensions.length > 0) {
-        if (!extra.extraFileExtensions.includes(fileExtension)) {
-          errorLines.push(
-            `${nonStandardExt}. It should be added to your existing "parserOptions.extraFileExtensions".`,
-          );
-          hasMatchedAnError = true;
-        }
-      } else {
-        errorLines.push(
-          `${nonStandardExt}. You should add "parserOptions.extraFileExtensions" to your config.`,
-        );
-        hasMatchedAnError = true;
-      }
-    }
-
-    if (!hasMatchedAnError) {
-      errorLines.push(
-        'The file must be included in at least one of the projects provided.',
-      );
-      hasMatchedAnError = true;
-    }
-
-    throw new Error(errorLines.join('\n'));
-  }
-
-  return astAndProgram;
-}
-
-/**
- * @param code The code of the file being linted
- * @param options The config object
- * @returns If found, returns the source file corresponding to the code and the containing program
- */
-function getASTAndDefaultProject(
-  code: string,
-  options: TSESTreeOptions,
-): ASTAndProgram | undefined {
-  log(
-    'Attempting to get AST from the default project(s): %s',
-    options.filePath,
-  );
-
-  const fileName = options.filePath || getFileName(options);
-  const program = createProgram(code, fileName, extra);
-  const ast = program && program.getSourceFile(fileName);
-  return ast && { ast, program };
-}
-
-/**
- * @param code The code of the file being linted
- * @returns Returns a new source file and program corresponding to the linted code
- */
-function createNewProgram(code: string): ASTAndProgram {
-  log('Getting AST without type information');
-
-  const FILENAME = getFileName(extra);
-
-  const compilerHost: ts.CompilerHost = {
-    fileExists() {
-      return true;
-    },
-    getCanonicalFileName() {
-      return FILENAME;
-    },
-    getCurrentDirectory() {
-      return '';
-    },
-    getDirectories() {
-      return [];
-    },
-    getDefaultLibFileName() {
-      return 'lib.d.ts';
-    },
-
-    // TODO: Support Windows CRLF
-    getNewLine() {
-      return '\n';
-    },
-    getSourceFile(filename: string) {
-      return ts.createSourceFile(filename, code, ts.ScriptTarget.Latest, true);
-    },
-    readFile() {
-      return undefined;
-    },
-    useCaseSensitiveFileNames() {
-      return true;
-    },
-    writeFile() {
-      return null;
-    },
-  };
-
-  const program = ts.createProgram(
-    [FILENAME],
-    {
-      noResolve: true,
-      target: ts.ScriptTarget.Latest,
-      jsx: extra.jsx ? ts.JsxEmit.Preserve : undefined,
-      ...defaultCompilerOptions,
-    },
-    compilerHost,
-  );
-
-  const ast = program.getSourceFile(FILENAME)!;
-
-  return { ast, program };
-}
-
-/**
- * @param code The code of the file being linted
- * @param options The config object
- * @param shouldProvideParserServices True iff the program should be attempted to be calculated from provided tsconfig files
- * @returns Returns a source file and program corresponding to the linted code
- */
-function getProgramAndAST(
-  code: string,
-  options: TSESTreeOptions,
-  shouldProvideParserServices: boolean,
-  createDefaultProgram: boolean,
-): ASTAndProgram | undefined {
-  return (
-    (shouldProvideParserServices &&
-      getASTFromProject(code, options, createDefaultProgram)) ||
-    (shouldProvideParserServices &&
-      createDefaultProgram &&
-      getASTAndDefaultProject(code, options)) ||
-    createNewProgram(code)
-  );
-}
-
 function applyParserOptionsToExtra(options: TSESTreeOptions): void {
   /**
    * Turn on/off filesystem watchers
@@ -259,12 +113,14 @@ function applyParserOptionsToExtra(options: TSESTreeOptions): void {
    */
   extra.range = typeof options.range === 'boolean' && options.range;
   extra.loc = typeof options.loc === 'boolean' && options.loc;
+
   /**
    * Track tokens in the AST
    */
   if (typeof options.tokens === 'boolean' && options.tokens) {
     extra.tokens = [];
   }
+
   /**
    * Track comments in the AST
    */
@@ -272,12 +128,23 @@ function applyParserOptionsToExtra(options: TSESTreeOptions): void {
     extra.comment = true;
     extra.comments = [];
   }
+
   /**
    * Enable JSX - note the applicable file extension is still required
    */
   if (typeof options.jsx === 'boolean' && options.jsx) {
     extra.jsx = true;
   }
+
+  /**
+   * Get the file extension
+   */
+  if (typeof options.filePath === 'string') {
+    extra.filePath = options.filePath;
+  } else {
+    extra.filePath = getFileName(extra);
+  }
+
   /**
    * The JSX AST changed the node type for string literals
    * inside a JSX Element from `Literal` to `JSXText`.
@@ -288,6 +155,7 @@ function applyParserOptionsToExtra(options: TSESTreeOptions): void {
   if (typeof options.useJSXTextNode === 'boolean' && options.useJSXTextNode) {
     extra.useJSXTextNode = true;
   }
+
   /**
    * Allow the user to cause the parser to error if it encounters an unknown AST Node Type
    * (used in testing)
@@ -298,6 +166,7 @@ function applyParserOptionsToExtra(options: TSESTreeOptions): void {
   ) {
     extra.errorOnUnknownASTType = true;
   }
+
   /**
    * Allow the user to override the function used for logging
    */
@@ -341,6 +210,7 @@ function applyParserOptionsToExtra(options: TSESTreeOptions): void {
   ) {
     extra.extraFileExtensions = options.extraFileExtensions;
   }
+
   /**
    * Allow the user to enable or disable the preservation of the AST node maps
    * during the conversion process.
@@ -380,12 +250,12 @@ function warnAboutTSVersion(): void {
 // Parser
 //------------------------------------------------------------------------------
 
-export type AST<T extends TSESTreeOptions> = TSESTree.Program &
+type AST<T extends TSESTreeOptions> = TSESTree.Program &
   (T['range'] extends true ? { range: [number, number] } : {}) &
   (T['tokens'] extends true ? { tokens: TSESTree.Token[] } : {}) &
   (T['comment'] extends true ? { comments: TSESTree.Comment[] } : {});
 
-export interface ParseAndGenerateServicesResult<T extends TSESTreeOptions> {
+interface ParseAndGenerateServicesResult<T extends TSESTreeOptions> {
   ast: AST<T>;
   services: ParserServices;
 }
@@ -394,9 +264,9 @@ export interface ParseAndGenerateServicesResult<T extends TSESTreeOptions> {
 // Public
 //------------------------------------------------------------------------------
 
-export const version: string = require('../package.json').version;
+const version: string = require('../package.json').version;
 
-export function parse<T extends TSESTreeOptions = TSESTreeOptions>(
+function parse<T extends TSESTreeOptions = TSESTreeOptions>(
   code: string,
   options?: T,
 ): AST<T> {
@@ -436,12 +306,7 @@ export function parse<T extends TSESTreeOptions = TSESTreeOptions>(
    * Create a ts.SourceFile directly, no ts.Program is needed for a simple
    * parse
    */
-  const ast = ts.createSourceFile(
-    getFileName(extra),
-    code,
-    ts.ScriptTarget.Latest,
-    /* setParentNodes */ true,
-  );
+  const ast = createSourceFile(code, extra);
 
   /**
    * Convert the TypeScript AST to an ESTree-compatible one
@@ -450,9 +315,10 @@ export function parse<T extends TSESTreeOptions = TSESTreeOptions>(
   return estree as AST<T>;
 }
 
-export function parseAndGenerateServices<
-  T extends TSESTreeOptions = TSESTreeOptions
->(code: string, options: T): ParseAndGenerateServicesResult<T> {
+function parseAndGenerateServices<T extends TSESTreeOptions = TSESTreeOptions>(
+  code: string,
+  options: T,
+): ParseAndGenerateServicesResult<T> {
   /**
    * Reset the parse configuration
    */
@@ -491,7 +357,6 @@ export function parseAndGenerateServices<
     extra.projects && extra.projects.length > 0;
   const { ast, program } = getProgramAndAST(
     code,
-    options,
     shouldProvideParserServices,
     extra.createDefaultProgram,
   )!;
@@ -541,6 +406,14 @@ export function parseAndGenerateServices<
   };
 }
 
-export { TSESTreeOptions, ParserServices };
+export {
+  AST,
+  parse,
+  parseAndGenerateServices,
+  ParseAndGenerateServicesResult,
+  ParserServices,
+  TSESTreeOptions,
+  version,
+};
 export * from './ts-estree';
-export { clearCaches } from './tsconfig-parser';
+export { clearCaches } from './create-program/createWatchProgram';
