+++ packages/typescript-estree/tests/lib/semanticInfo.ts
@@ -9,11 +9,11 @@ import {
   parseCodeAndGenerateServices,
 } from '../../tools/test-utils';
 import {
+  clearCaches,
   parseAndGenerateServices,
   ParseAndGenerateServicesResult,
 } from '../../src/parser';
 import { TSESTree } from '../../src/ts-estree';
-import { clearCaches } from '../../src/tsconfig-parser';
 
 const FIXTURES_DIR = './tests/fixtures/semanticInfo';
 const testFiles = glob.sync(`${FIXTURES_DIR}/**/*.src.ts`);
